+++ lib/nikfemm/src/algebra/coo.hpp
@@ -12,7 +12,7 @@ namespace nikfemm {
         uint32_t m = 0;  // rows
         uint32_t n = 0;  // columns
 
-        MatCOO(uint64_t m, uint64_t n);
+        MatCOO(uint32_t m, uint32_t n);
         ~MatCOO();
 
         void set_elem(uint32_t _m, uint32_t _n, double val);  // row, column, value
